まずはPingのみre-architect
