+++ api/cmd/yonda/main.go
@@ -1,0 +1,6 @@
+package main
+
+func main() {
+	Init()
+	Start()
+}
