+++ api/internal/domain/entity/book_details.go
