+++ api/internal/domain/entity/errors.go
