+++ api/internal/interface/actions/book_detail.go
@@ -2,7 +2,7 @@ package actions
 
 import (
 	"github.com/jinzhu/gorm"
-	"github.com/k1rnt/yonda/api/domain/entity"
+	"github.com/k1rnt/yonda/api/internal/domain/entity"
 	"github.com/labstack/echo/v4"
 	"net/http"
 )
