--- api/main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/k1rnt/yonda/api/router"
-)
-
-func main() {
-	router.Init()
-	router.Start()
-}
